+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # CHANGELOG
 
-## 1.2.0
+# v0.1.0-alpha
+ 
+## What's new
+- Added droppable items for package and species
+  - These will set initial attributes
+  - Initial skills
+  - If a skill is granted by package and species it will increase the rank by 1
+- Hide cost from skills, and non-gear items
+- Added a skill max (rank 4)
+- Make the skills for species and package an area to drop skill items instead of a text field
 
-- Add support for Foundry v10
+## Bugs
+- Ranged mods on a weapon attack are gone - repair
+- Fix status toggles stunned gets stuck
+- Apply damage is applying damage to target and selected token - add selector to confirmation dialog
+- Degree of success on weapon attacks is gone - restored